Moved common attribute visitors to global resources
