+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/attributes/resources/DeviceInfoVisitor.java
@@ -16,18 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.traces.common.attributes;
+package co.elastic.apm.android.sdk.attributes.resources;
+
+import android.os.Build;
 
 import co.elastic.apm.android.sdk.attributes.AttributesVisitor;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.semconv.resource.attributes.ResourceAttributes;
 
-public class RuntimeDescriptorVisitor implements AttributesVisitor {
+public class DeviceInfoVisitor implements AttributesVisitor {
 
-    @SuppressWarnings("ConstantConditions")
     @Override
     public void visit(AttributesBuilder builder) {
-        builder.put(ResourceAttributes.PROCESS_RUNTIME_NAME, "Android Runtime")
-                .put(ResourceAttributes.PROCESS_RUNTIME_VERSION, System.getProperty("java.vm.version"));
+        builder.put(ResourceAttributes.DEVICE_MODEL_IDENTIFIER, Build.MODEL)
+                .put(ResourceAttributes.DEVICE_MANUFACTURER, Build.MANUFACTURER);
     }
 }
