+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/attributes/resources/OsDescriptorVisitor.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.traces.common.attributes;
+package co.elastic.apm.android.sdk.attributes.resources;
 
 import android.os.Build;
 
@@ -24,11 +24,23 @@ import co.elastic.apm.android.sdk.attributes.AttributesVisitor;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.semconv.resource.attributes.ResourceAttributes;
 
-public class DeviceInfoVisitor implements AttributesVisitor {
+public class OsDescriptorVisitor implements AttributesVisitor {
 
     @Override
     public void visit(AttributesBuilder builder) {
-        builder.put(ResourceAttributes.DEVICE_MODEL_IDENTIFIER, Build.MODEL)
-                .put(ResourceAttributes.DEVICE_MANUFACTURER, Build.MANUFACTURER);
+        builder.put(ResourceAttributes.OS_DESCRIPTION, getOsDescription())
+                .put(ResourceAttributes.OS_VERSION, Build.VERSION.RELEASE)
+                .put(ResourceAttributes.OS_NAME, "Android");
+    }
+
+    private String getOsDescription() {
+        StringBuilder descriptionBuilder = new StringBuilder();
+        descriptionBuilder.append("Android ");
+        descriptionBuilder.append(Build.VERSION.RELEASE);
+        descriptionBuilder.append(", API level ");
+        descriptionBuilder.append(Build.VERSION.SDK_INT);
+        descriptionBuilder.append(", BUILD ");
+        descriptionBuilder.append(Build.VERSION.INCREMENTAL);
+        return descriptionBuilder.toString();
     }
 }
